Update index_storage.txt
I hadn't put the folder with where the user notes are stored in gitignore, but since deleting files locally doesn't delete those files in github, I have manually deleted all but one of the directories (read as `labels`) left as an example. I updated this index file so that it doesn't refer to files that don't exist.